Update README.md
According to various research, the average global temperature has been rising over the last century. An increase in the global temperature that doesn't stop would have disastrous effects. There will be an impact on billions of people due to rising sea levels and an increase in the frequency of severe weather occurrences.

Question 1: Could it be that certain airborne gases are causing the environment to get overheated?

Question 2: If so, which gas? What is the relationship between the gases and the heat that is trapped and causes the temperature to rise?